Removed excessively defensive code
The code previously checked a number of assumptions that should never
actually be false. These are likely left-over from before Content Types
were introduced, when topic objects were created the first time they
were referenced, and otherwise loaded. That is no longer the case, so a
lot of the legacy code shouldn't be needed.