Marked Topic(string key) constructor as protected
Topics should always be created with, at minimum, a Key and a
ContentType. For creating topics, this should be done via the Create()
factory method to ensure that the type maps to the ContentType if a
strongly typed derived class is available. That said, derived classes
will already know the ContentType, so this isn't an issue. For these
scenarios, the protected Topic(string key) constructor will infer the
Content Type via GetType().