Established Repository Classes
The Repository classes largely duplicate the existing Provider classes.
They primarily vary in that a) overloads are greatly reduced, b)
optional parameters are prefered to overloads, and c) most types no
longer have return types (e.g., Move is void, not bool). If an error
occurs, an exception should be thrown, not a status code.