Remove dependencies on TopicRepository.ContentTypes
Previously, SqlTopicRepository (and SqlTopicDataProvider before it) were
dependent on TopicRepository.ContentTypes to determine whether a topic
should be stored as indexed or in the blob. This makes it harder to
write test cases for, less flexible in working with separate data sets,
and will prevent it from being used with Dependency Injection (since
TopicRepository is a hard-coded dependency). To mitigate this, it has
been updated to maintain its own internal list of Content Types, or use
itself to retrieve a list of Content Types as needed.