Remove TopicRepository dependency
Topic, as an entity, should not have dependencies on the data repository
(ITopicRepository). To remedy this, lazy loading is removed (previously
moved to e.g. SqlTopicRepository) and CRUD convenience passthroughs are
removed (they should be called off ITopicRepository now). In addition,
the invasive portion of Parent (that performs the actual Move() was
migrated to ITopicRepository (in the TopicRepositoryBase).