Removed legacy convenience CRUD calls on Topic
E.g., replaced topic.Save() with TopicRepository.DataProvider.Save().
This still relies on the Service Locator (anti-pattern) but this is
legacy ASP.NET WebForms code that will be shuffled off to a separate DLL
for backward compatibility only, that is acceptable. (All new code and
sites will use MVC with dependency injection).