Removed Provider model
Removed the legacy ASP.NET Provider model from the topic configuration,
including the TopicDataProviderManager that exposed it. Now, new
instances of the ITopicProvider must be manually instantiated in the
Composition Root. For legacy ASP.NET Web Forms clients, the
ITopicProvider should be instantiated in the global.asax and set to the
TopicRepository.DataProvider for backward compatibility.