Introduced connection string dependency
Previously, the SqlTopicRepository implemented a hard-coded dependency
on the ConfigurationManager in order to retrieve a connection string. It
can now receive one via the constructor, thus relying on dependency
injection instead of a hard-coded dependency. This is made possible by
removing the legacy Provider Model, which necessitated that any Provider
have a null constructor (and, thus, didn't permit dependency injection).