Replaced RouteData with TopicRoutingService
Previously, routing data was relayed from TopicsRouteHandler to
TopicPage to help reduce the amount of information it needed to
recalculate. Given the relative effeciency of the TopicRoutingService,
however, this provides a more direct approach to retrieving the
PageTopic without relying on RouteData. This also reduces the subtle
dependencies between the two services.