Fixed IsValidView and ViewPath
Previously, IsValidView was always returning false due to a logic
error--and, even if it didn't, it would have returned a path that was
incompatible with the ViewPath property (i.e., it would have included
the ContentType twice). Resolving these issues allows the View overrides
via e.g., QueryString, and should work with headers as well.