Removed unused SQL Tests
We _should_ have a unit test project for the SqlTopicRepository. This will, however, require that the SqlTopicRepository be refactored so that the dependency on SQL server can be replaced with a stub. Until then, there is little benefit or use to having a test project, and it provides the misconception that there are corresponding unit tests available (not actually true).