Established RelationshipFacade
The RelationshipFacade provides a simplified wrapper for abstracting the internal dynamics of the Relationship storage, forcing the user to go through a more reliable and foolproof interface, instead of needing to be aware of how the related topics are stored internally. Additionally exposes the ability to filter incoming relationships by namespace (acknowledging these may be compressed if multiple ContentTypes have the same relationship scope).