Implemented infinite loop handling
`SetValue()` will look for a local property and, if it exists, call it, as a means of enforcing local business logic. This means local properties must tell `SetValue()` when they're calling it to avoid an infinite redirect. To help avoid this scenario, a runtime error will occur when `SetValue()` is called more than three times in a row.