Made `AttributeValue` immutable
Made `AttributeValue` immutable except for `IsDirty` and the new `EnforceBusinessLogic` state properties. `EnforceBusinessLogic` is now used in place of the previous "__key" convention, which was not only confusing, but could, in rare scenarios, potentially cause bugs (e.g., if a thread tries to access "key" between `SetValue()` and `InsertItem`, during which point it would be set to "__key").