Introduced new `AttributeSetter` Attribute
The `AttributeSetter` attribute marks a property as being a setter for an `AttributeValue`; if a property is decorated with this, then it will be called when `AttributeValueCollection.SetValue()` is called assuming a) the names match, and b) the `AttributeValue.EnforceBusinessLogic` property is set to `true`.

This negates the need for the previous `SettableTypes` and `BypassBusinessLogicList` properties, which maintained which types and attributes, respectively, were eligible for enforcement.