Removed dependency on `Uri`, `RequestContext`
`TopicRoutingService` no longer derives from `Uri`, thus allowing it to expose only those members defined in the new `ITopicRoutingService` interface. In addition, while it still _accepts_ `RequestContext`, it can _instead_ be provided directly with the `RouteValueDictionary` and `NameValueCollection`, thus making it more portable. (Though the `RequestContext` dependency will likely remain an issue for ASP.NET Standard migration, so that remains a concern for the future).