Implemented custom `TopicViewResult`
The `TopicViewResult` a) requires the `TopicViewModel` (or a derivative), and b) contains logic for searching through appropriate views (e.g., query string, headers, topic, &c.). This largely removes the dependency on the `ITopicRoutingService` for finding the view, and thus it can be greatly simplified. Now returning `TopicViewResult()` by default from `TopicController`.