Split `TopicRoutingService` with `ViewRoutingService`
`TopicRoutingService` is no longer needed to provide view logic for the MVC app, due to the new `TopicViewEngine` and `TopicViewResult` classes. This allows the `TopicRoutingService to be greatly simplified. The legacy Web Forms project, however, still needs parsing of the views folder. As such, the views logic was separated into a new `ViewRoutingService` which derives from the `TopicRoutingService`, thus maintaining backward compatibility.