Moved `ITopicRepository` contracts to `ITopicRepositoryContract`
Previously, the contract class was bound to the `TopicRepositoryBase` and only covered the `abstract` members. This created an inconsistent application of code contracts (some defined in `TopicRepositoryBaseContract`, others defined in `TopicRepositoryBase`). It also resulted in warnings due to concrete implementations inheriting from both `ITopicRepository` and `TopicRepositoryBase`. Moving the contracts over to `ITopicRepositoryContract` and applying `ITopicRepository` to the `TopicRepositoryBase` mitigates this. It's also much cleaner code.