Established explicit conversions
When saving a `TopicId` to the `topics_AttributeValues` table, the value is being implicitly converted from an `int` to a `varchar(255)`, and vice versa. This could result in data loss, so SQL's code analysis throws a warning. This comes up when saving e.g., `TopicID` (`DerivedTopic`) and `ParentId` (`Parent`). To mitigate this, am providing an explicit conversion (e.g., `CONVERT(VarChar(255), @TopicID)`).