Allowed "true" for booleans
Boolean values are, by convention, stored as either a "1" or a "0" in the database. To provide a bit more flexibility, the `SetProperty()` method will now also accept "true" (which will get converted to a "1" by the corresponding property). The data still must be saved as a "1" in the database, but this makes the setting of properties a bit more fool-proof.

(That said, this could cause confusion since a person calling `SetValue("Boolean", "true")` would expect `GetValue("Boolean").Equals("true")`; instead, they will find that  `GetValue("Boolean").Equals("1")`).