Prevented duplicate key scenario
By using a `TopicCollection` for the return type of `FindAllByAttributeName()`, `Topic` introduces the potential to return multiple topics with the same key in the collection. To mitigate this issue, duplicates are not permitted. (This could potentially orphan desired results; to mitigate that issue, though, a different data structure would need to be used.)