[Incomplete] Scaffolding for `ITopicMappingService`
The current version supports:
1. Finding `{ContentType}TopicViewModel` in the current assembly (if not specified),
2. Adding scalar values (from `Attributes.GetValue()`)
3. Crawling parents

Other scenarios, such as `Children`, `Related*`, and nested topics are scaffolded, but not yet tested, and likely need additional work. Further, other scalar value types should be evaluated, such as `DateTime`, `Float`, &c. (Currently, it only supports `Bool`, `Int`, and `String`).

Finally, this needs to be modified to be able to find `{ContentType}TopicViewModel` instances in other assemblies, namespaces. (This is also a requirement that needs to be investigated for the `TopicFactory.Create()` method.)