Initial TopicViewModel implementations
While the TopicViewModels should be implemented in the presentation layer, the Topic Library will ship with a set of default implementations for convenience. Implementers can use these directly, extend them, or ignore them. They represent the out-of-the-box content types with their factory default configurations (acknowledging that customers may extend these definitions as well).