Instructed to crawl all assembles
The updated `GetViewModelType()` method will now search _all_ assemblies (not just the _current_ assembly) and find any classes that end with `{ContentType}TopicViewModel` _regardless_ of which namespace they're in. This is really slow (around 1s) but should only need to be run once per application lifetime; as a result, a cache is maintained of all classes matching the convention.