Updated interface to include `includeRelationships`
By default, mapping will follow the object graph, including parents, children, relationships, and nested topics. If `includeRelationships` is true it will populate all of these. If false, it will only follow parents and nested topics. This is automatically set to false for any relationships that are followed; that is, it is not recursive. For instance, if a relationship is followed, the relationships of the relationship will not be.