Established an `ITopicViewModelCore`.
The `TopicMappingService` does not mandate any interfaces or base classes in order to map a `Topic` to a DTO. That said, certain presentation implementations will find it useful to ensure access to certain base properties, such as `Key`, `ContentType`, and `View`. To facilitate those potential needs, while not locking consumers into the default `TopicViewModel` implementation, the `ITopicViewModelCore` provides flexibility for those any presentation libraries that mandate a baseline set of properties.