Added support for Getter Methods
Modified `TopicMappingService` so that it will default to a `Topic.Get{Property}()` method, if it exists, before calling `Topic.Attributes.GetValue({Property})`. This allows dynamically assessed values, such as `GetWebPath()` and `GetUniqueKey()` to be executed and their results stored on the `TopicViewModel`. These could have been recreated on `TopicViewModel`, but that would have mandated all descendants always be present (up to the root)—which won't be true for e.g. related topics. Further, this approach provides flexibility for other dynamically determined values in the future. It assumes that the type can be converted to a string, is not generic, and has no parameters.