Ensure collections are sorted
By default, all collections (e.g., Children) are sorted when loaded from the repository. But if their sort order is changed in the editor, that change won't be fully reflected until the data is reloaded. This is because it's involved to change the sort order of an existing collection. To ensure this doesn't impact views, calling `Sorted` on `Children` references ensures that they are correctly sorted.