Simplified `FakeTopicRepository`
Previously, the `FakeTopicRepository` managed its own cache, and reproduced lookups. This works is already done by `CachedTopicRepository`. By using `CachedTopicRepository`, we avoid duplicating unnecessary code—and additionally get the benefit of testing actual production code. Negatively, this means `CachedTopicRepository` becomes implicated in the `ITopicRepositoryTest` and `TopicRoutingServiceTest` cases, but it's a lightweight façade, so this isn't a big issue—and especially if its code was going to be duplicated anyway.