Established `OnActionExecuting` validation filter
Previously, the `CurrentTopic` was validated as part of the `Index()` method. This made `TopicController` difficult to subclass since all of this validation would need to be duplicated in any derived controllers. By moving this logic to the `OnActionExecuting()` event handler, this logic is enforced on _all_ actions on either `TopicController` or any derived class.