Introduced `TopicEntityViewModel`
In some scenarios, a view may require full access to an `ITopicRepository` as well as a `Topic` object graph. In these cases, mapping a partial topic tree to topic view models (e.g., with the `ITopicMappingService`) is an extra level of unnecessary overhead. In these cases, the `TopicEntityViewModel` provides support for low-level access to the entity without going through a topic view model.