Introduced reusable `SitemapController`
Technically, the `SitemapController` could be used for _any_ `TopicEntityViewModel`, and we may consider renaming it to a more generic use case (e.g., `TopicGraphController`). In the meanwhile, however, it effectively provides a foundation for any view which requires access to the entire object graph via a `TopicEntityViewModel`.

Be aware that the unit test for this does not perform full validation due to a dependency on a missing `ControllerContext`; in absence of that, the expected exception is confirmed.