Fixed issue with default identifiers
The `FakeRepository` was updated to assign `Id`s to the `Root:Web` and its children, in order to support the `LayoutController_Menu` test. This disrupted the `ITopicRepositoryTest_SaveTest` test. To mitigate this, the latter was updated to test the `Id`s of the `Root:Configuration` instead (which retains the default values of `-1`).