Made generic controllers abstract
Generic controllers cannot be effectively routed to by ASP.NET MVC. To mitigate this, a non-generic derived type must be created that sets the generic type. To help enforce this, these classes have been marked as `abstract`. Further, they've been renamed to `*ControllerBase` to communicate that they're not meant as concrete implementations.

This required creating concrete stubs to ensure that the unit tests continue to work.