Removed empty `Topic` default
Previously, the `Redirect()` method defaulted to creating a `new Topic()`, which would avoid an exception, but only by providing a nonsensical default (i.e., a redirect to the homepage, since an empty `Topic` will have a `GetWebPath()` of `/`). Removing this will result in an exception, which will accurately communicate an unexpected state.