Remove `SortOrder`
`SortOrder` is a legacy concept due to the fact that when a `Parent` was changed, the `Topic` was always placed at the end of the `Topic.Children` collection. This is mitigated by the new `SetParent()` method which allows a `Topic` to be moved in the in-memory object graph while honoring the specific location relative to its siblings.