Established `Topic(string, string, Topic, int)` constructor
Previously, `Topic` had a default constructor, which the `TopicFactory` used to create new instances. Updated the constructor to require, at minimum, `key` and `contentType` and, optionally, `parent`, and `id`. This ensures there isn't a means of creating an "empty" (invalid) `Topic`. Updated derived classes and `TopicFactory` to honor this (which greatly simplifies the latter).