Generalized method parameters as `IList<Topic>`
Previously, `ReadOnlyTopicCollection` and `ReadOnlyTopicCollection<T>` had to be constructed explicitly using a `TopicCollection` or `TopicCollection<T>`. To make them more flexible, modified the behavior to accept an `IList<Topic>`. If the `IList<Topic>` is a `TopicCollection<T>` then the underlying collection uses it as is for keying; otherwise, it is converted to a `TopicCollection<T>`. This provides more flexibility, while not changing the normal day-to-day usage of the methods.