Prevent `Move()` on new `Topic`
Currently, `Move()` is called during `Save()` if the `ParentId` has changed prior to the last `Save()` or `Load()`. This is a reasonable assumption for existing topics. For new topics, however, it leads to a `Topic` being moved in the database even though that `Topic` doesn't yet exist. To mitigate this, put in an exception to check the status of the topic.