Fixed string conversion
It is fully expected that `GetMethodValue()` might return an `int` or `bool` or `DateTime`. With `as string` those would be treated as `null`. Instead, converted to `ToString()`. Negatively, this means reference objects might be converted to their default `ToString()` behavior—but that should be an outside case where there is a `Get{Property}()` method that the developer does not intend to call and which returns an incompatible content type.