Introduced support for 1:1 property matching
Previousy, if there was a property on topic named the same value as a property on the target DTO, the mapping library wouldn't bind them. This is because it was primarily looking for e.g. `Attribute.GetValue()` mappings. This new addition provides support for 1:1 property matching—at least assuming `String`, `Int`, `DateTime`, or `Bool`. This also eliminates the need for the `Id` fix (since `Id` doesn't have an attribute, but can now be mapped using the generic `GetPropertyValue()` technique).