Renamed `[Recurse()]` to `[Follow()]`
This is more intuitive since the `[Recurse()]` attribute does not instruct the `ITopicMappingService` to be _truly_ recursive; it only works that way if the attribute is _chained_. The semantics of `[Follow()]` make this more clear.

While I was at it, I also refactored the `Relationships` property to use an automatic implementation.