Moved attribute settings into `PropertyConfiguration` class
This technically increases the code significantly, but almost all of that is in documentation. In exchange, not only does this improve the documentation by complementing the documentation on each attribute, but it also makes it easier to reuse those attributes in other potential implementations of the `ITopicMappingService`. Finally, it reduces the code slightly in the `TopicMappingService` implementation.