Exposed `TypeLookupService` to `TopicFactory`
By default, `TopicFactory` will use the local default of `TopicLookupService` for identifying strongly-typed `Topic` classes. Optionally, this can be overwritten, thus allowing an alternate source of `Topic` classes. As part of this, also introduced the new `FakeTopicLookupService` in the unit tests, and configured the `AssemblyInit()` to initialize it when the test assembly loads. This prevents the unit tests from needing to use any reflection, or rely on dynamic lookup of the classes, while simultaneously reducing the number of (production) classes involved in each test.