Introduced `DefaultTopicLookupService`
This is a custom configuration of the `StaticTypeLookupService` that ensures that the basic classes (such as `ContentTypeDescriptor`) are accounted for. These are necessary for backend support and may not be obvious if an unfamiliar developer attempts to establish a new `StaticTypeLookupService` from scratch. As such, this provides a less error-prone default implementation.