[Breaking] Converted `ITopicMappingService` to `async`
By converting `ITopicMappingService` to be `async`, we allow nested calls to `MapAsync()` to be done in parallel. This is especially useful when mapping related topics from collections; being able to do those in parallel should be considerably faster.

This mandates that all calling items—such as the unit tests and MVC controllers—also be made async. Updated accordingly.