Bug Fix: Second cache for null type
The second cache attempt is intended to be for the default implementation—i.e., one that is of the type `{ContentType}TopicViewModel`—and to be used when the type is not explicitly specified via e.g. `MapAsync<T>(…)`. It was using the same cache key as the first cache, however. This would actually work in the majority of cases, but to extend it's usefulness this creates a new cache key with a null type for the second cache. This allows the cache to be set even when `MapAsync<T>()` is used to explicitly request the default mapping (otherwise, the first would have been used with `MapAsync<T>()` and the second would have been used with `MapAsync()`, potentially requiring two mappings of the same object). Because the caching logic is complex, also extended the XmlDoc.