Added concurrency to `AddNestedTopicsAsync`
`AddNestedTopicsAsync()` now implements a local `taskQueue` so that it can process multiple children simultaneously. It also implements a lock pattern so that those children are only added to the source collection once they're fully populated; this avoids potential race conditions with concurrency where duplicates could be added to the cached topic.