Fixed partial match bug in `IsSelected`
Previously, `IsSelected()` would mark `/Web/Cat` as being selected if the user was on `/Web/Catacomb` because it relied on a `StartsWith()` query against `UniqueKey`. Fixed this by appending `:` to both the source and the target to ensure that the last `Key` in the `UniqueKey` was a full match, not a partial match.