Simplified test names
Removed the class name prefix (since it's listed in the class name, and tests can be grouped by class) and removed the "Test" suffix (since these are all test methods in test classes, this should be implicit). We may revisit these again in the future, but for now they're more succinct and easier to read in Visual Studio's Test Explorer.