Throw error on duplicate
Under expected circumstances, a duplicate `MemberInfo` key should not exist. This will happen, however, if the `new` keyword is used on a member of the _current_ class (as opposed to a _derived_ class). In that case, an error will occur. This should not happen; if it does, the developer is warned about the duplicate key with a specific error message.