Initialized `_cache_ as part of constructor
Whenever a `CachedTopicRepository` is called, it checked to see if the `_cache` was initialized. This should not be necessary. If a `CachedTopicRepository` is required, then it should immediately load its cache so it's ready for the first request. In most scenarios, this won't change the load time (as the `Load()` method will be called shortly after establishing the cache), but it does simplify the internal logic.