Established `GetContentTypeDescriptors()` test
Ensured that `ITopicRepositoryTest` is evaluating the `GetContentTypeDescriptors()` method. This is important since this method has some built-in logic.

As part of this, also ensured that `FakeTopicRepository`'s `Load()` and `GetContentTypeDescriptors()` methods were implemented enough to satisfy the test requirements.

This included registering all content types in the `FakeTopicRepository` to ensure that internally-created `ContentTypeDescriptor`s had been created (e.g., `Lookup`, `List`, &c.) as well as renaming `ContentType` to `ContentTypeDescriptor`.