Moved to `catch(SqlException…)`
Previously, we were catching `Exception`, as per Microsoft's guidance. This was due to the fact that there are a number of other exceptions that can be raised during the processing of a `SqlConnection` and, especially, a `SqlDataReader`. Most of those, however, are not _expected_ exceptions, and thus should not be handled (or even wrapped in a generic exception).

In addition, moved `Exception` messages to use string interpolation for cleaner code.