Introduced `TopicRepositoryException`
Previously, the `SqlTopicRepository` was throwing generic `Exception`s. This is not a best practice. It is appropriate, however, for implementations of `ITopicRepository` to throw a generic, database agnostic exception (as opposed to e.g. `SqlException`) since it should not be tied to any specific persistence technology. The `TopicRepositoryException` allows it to accomplish this.