Provide more explicit messaging on key conflict
When a `KeyedDictionary` finds a duplicate key, it provides a generic `ArgumentException` announcing this. It doesn't provide any context, however, such as the key that is being duplicated. This makes debugging difficult. To mitigate this, I've added explicit `InsertItem` overrides that detect duplicate keys and, if they exist, provide an `ArgumentException` with an explicit error message containing not only the key value, but also any other relevant context (such as any associated `Topic` or `T` type, as appropriate).