Replaced references to contracts library
Replaced references to `System.Diagnostics.Contracts` with references to `Ignia.Topics.Diagnostics`. The latter provides a partial reproduction of the former's interface with the goal of maintaining syntactical consistency—and, in the case of `Contract.Requires()`, basic functional consistency. This is a stopgap measure to remove dependencies on Microsoft's deprecated Code Contracts project.