Distributed logic from `ITopicRepositoryContract`
With Code Contracts removed, the interface contract classes—as specified by the `[ContractFor]` attribute—no longer function. As a result, the logic from this class has been distributed as appropriate. Most of it was able to be moved to the `TopicRepositoryBase` class. For the `Load()` method, however, the validation needed to be moved to each concrete implementation since there isn't a base `Load()` method to derive from—nor does it make much sense to otherwise have one.