Removed unnecessary `Ensures()`
Due to the removal of Code Contracts, the `Ensures()` method no longer serves a function. In addition, however, it is not strictly needed in many cases, where there's no expected scenario where a method would return e.g. a null value (e.g., if a new instance is created inside the method, or an instance is created by a method that itself ensures the result is not null). In these cases, the use of `Ensures()` was primarily intended to satisfy the Code Contracts analysis module, and can simply be removed.