Replaced `Assume()` with `Requires<T>()` in `GetKeyForItem()` on `KeyedCollection`s
With Code Contracts removed, `Assumes()` is no longer necessary as part of the code analysis module and can be removed. That said, with `KeyedCollection`s, the `Assume()` call on `GetKeyForItem()` can be replaced with a `Requires<T>()` call. This isn't usually expected since the framework is the typical client and will always pass an `item`. Since `GetKeyForItem()` is `protected`, however, it could theoretically be called from an external source. As such, validating the parameters is a best practice.