Introduced custom `Assumes<T>()` method
The Code Contracts base implementation includes a non-generic `Assumes()` method which provides no runtime validation—it is exclusively used by static code analysis. The custom `Assumes<T>()` method is used in the same conditions—but, unlike `Assumes()`—actually provides runtime validation of the current state. In this way, it operates exactly like `Requires<T>()` but it syntactically represents validation of runtime state and not preconditions. As such, it can—and is expected to!— be placed within the body of a method, not exclusively at the beginning of a method.