Removed `Assume()` for `Parameters` collection
It should always be safe to assume that the `Command` object has a `Parameters` collection—even if the length is zero. While Code Contracts needed to be told this for the purposes of static analysis, our code doesn't require runtime validation to be assured that this is true. This assumption can be removed.