Implemented `Assume<T>()` on `Ignia.Topics.Web`
The default Code Contracts `Assume()` method is only used by static analysis. Now that Code Contracts is deprectated, this is no longer necessary. As such, references to `Assume()` have been removed—if, in fact, they can be safely assumed—or replaced with the custom `Assume<T>()` method—which provides actual runtime validation of state.  This was previously done to the rest of `Ignia.Topics`, but missed on `Ignia.Topics.Web` since it isn't subject to realtime validation in Visual Studio.