Marked the `Result<T>()` method as obsolete
The `Result<T>()` method is used with Code Contract's `Ensures()` to return the result of a method and verify that it returned the expected value. This requires Code Contract's rewriting module. Since that is deprecated, and `Ensures()` has been marked as obsolete, this no longer serves a purpose and can safely be marked as obsolete itself.