Established `AspNetCore` project
The `Ignia.Topics.AspNetCore.Mvc` project "forks" the previous `Ignia.Topics.Web.Mvc` project as an *ASP.NET Core 2.x* variation. The history of all files will be shared with the `Ignia.Topics.Web.Mvc` project. This honors the fact that much of the code should remain the same—and, thus, the history of how that code came to be will remain relevant. This initial commit updates the `csproj` file variables, swaps out legacy *ASP.NET MVC Framework* *NuGet* packages with *ASP.NET Core* *NuGet* packages, and updates the project's namespace. THIS COMMIT WILL NOT COMPILE; additional updates are necessary to get this project to a working state.