Corrected parsing of accept headers
In *ASP.NET Core*, individual headers are returned from `RequestContext` as a string, not a collection. As such, the first item does not need to be retrieved from the array, as it's already being returned as a string. (Accept headers are typically submitted as a comma-delimited list, as opposed to multiple accept headers.)