Established extension method for configuring OnTopic
The `ServiceCollectionExtensions.AddTopicSupport()` provides a convenient way to register *OnTopic* related features with the *ASP.NET Core* configuration pipeline. In this case, it adds  support for the new `TopicViewLocationExpander` so that it doesn't need to be configured manually by implementers. Additional items may be added to this in the future.