Resolved boolean expression mixups
The boolean expressions were not resolving correctly due to three issues. First, in the term `?? false &&` everything after the `&&` is only evaluated if the `??` is invoked (and, thus, `view` is `null`); the intent and functionality of this is clarified with parenthesis. Second, without parenthesis, counterintuitively, the `!` only applies to the expression prior to the `??`, thus necessitating e.g. `!view?.Success ?? true` to ensure a condition is tripped if `view` is `null`. And third, the last condition was just backwards—it should fire when `view.Success` is `true`, not `false` (whoops!).