Corrected paths and tokens
There were two issues with the paths. First, they should now begin with `/`, not `~/`, as they did in the previous version. Second, the order of `String.Format()` changed since the previous version, this `{1}` now refers to `Controller`, whereas it used to refer to `View`; to fix this, `{1}` was replaced with `{0}`, which now represents `View`.