Corrected reference to `RouteData`
Previously, the `RouteData` was being pulled from `RouteValues`. But this is actually only a _subset_ of `RouteData` represented _known_ route keys (e.g., `controller`, `action`). Custom route keys (e.g., `path` and, of relevance here, `contenttype`) are not maintained here. That said, the `RouteData` has the full set, and can be accessed directly off the `ActionContext` property. This returns as an `object`, though, so also requires some boxing to cast the variable as a `string` later in the code.