Introduced the `ExecuteResultAsync()` method
The `ExecuteResultAsync()` method was previously inheriting from the parent `ViewResult` class. The problem with this is that the method is responsible for calling the `IActionResultExecutor<>` service. As such, the `TopicViewResultExecutor` was never getting called. This is resolved by overriding this method—and _not_ calling the `base` version, instead executing our specific version of the `IActionResultExecutor`.