Addressed issue with `ViewData` dependency
In *ASP.NET Core*, the `ViewData` is no longer available, by default, to the `ViewResult` class. Instead, it is expected to be injected by callers—namely, `IController` instances. In `Controller`, this is done via the `Controller.View` command using property injection. Since `ViewResult` won't work correctly _without_ a populated version of `ViewData`, however, this should really be handled via constructor injection. As such, the constructor was updated (and streamlined) to require both `ViewData` and `TempData` to avoid confusion. Further, a `TopicView()` helper method was added to `TopicController` to add these parameters to a new instance of the `TopicViewResult`—alongside relaying the `topicViewModel` and, optionally, the `viewName` (which are parameters of the new `TopicView()` method). This is similar to how *ASP.NET Core*'s own `Controller` works, while addressing a design limitation of its own `ViewResult` which makes it easy to inadvertently introduce bugs (e.g., by returning a `new ViewResult()` while not setting the mandatory `ViewData` property).