Addressed potential locale conflicts
The library contained quite a few `Equals()`, `StartsWith()`, `Replace()`, and `ToString()` method calls without corresponding `CultureInfo` or `StringComparison` arguments providing guidance on how to handle different locale settings. This is fixed by either adding these or, in the case of `Equals()`, frequently replacing with the operator (`==`, `!=`) for the sake of brevity and clarity.