Suppress `CA1056` on View Models
`CA1056` is a warning that any string property ending in `Url` should be of type `Uri`. That's probably a good idea. But the `TopicMappingService` doesn't (yet) support this behavior and, thus, it's impractical for the view models library. This would be nice to have in the future. In the meanwhile, however, this warning is disabled from code analysis.