Supressed `CA1707` on tests
`CA1707` relates to the use of underscores in member names. This is a good rule, and consistent with Ignia's own style guide. It is, however, common to use underscores as delimeters in test names to make them read as natural language, or to separate out concepts (as we're doing here). As such, I'm supressing these warnings on the `Ignia.Topics.Tests` project.